+++ js/main.js
@@ -1,6 +1,6 @@
 import {ANNOUNCEMENT_COUNT} from './data.js';
 import {createAnnouncement} from './create-announcement.js';
-import {generationMarkupCards} from './generating-markup-cards.js';
+import {generateMarkupCards} from './generate-markup-cards.js';
 
 const map = document.querySelector('#map-canvas');
 
@@ -8,6 +8,6 @@ const similarAnnouncements =  new Array(ANNOUNCEMENT_COUNT)
   .fill('')
   .map((currentValue, index) => createAnnouncement(index));
 
-const markupCards = generationMarkupCards(similarAnnouncements);
+const markupCards = generateMarkupCards(similarAnnouncements);
 
-map.appendChild(markupCards.children[0]);
+map.appendChild(markupCards.children[1]);
